+++ js/data.js
@@ -1,0 +1,66 @@
+import {getRandomInteger, getRandomArrayElement, createIdGenerator} from './util.js';
+
+const PICTURE_COUNT = 25;
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const COMMENTS_MIN = 0;
+const COMMENTS_MAX = 30;
+const AVATAR_MIN = 1;
+const AVATAR_MAX = 6;
+
+const NAMES = [
+  'Иван',
+  'Александр',
+  'Сергей',
+  'Михаил',
+  'Алексей',
+  'Владимир',
+  'Антон',
+  'Дмитрий',
+  'Евгений',
+  'Николай'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTIONS = [
+  'Самая лучшая фотография',
+  'Самая профессиональная фотография',
+  'Вся красота мира в одной картинке',
+  'Моменты, которые запечатлены навсегда',
+  'Счастье в каждом кадре',
+  'Когда слова не нужны, достаточно фотографии',
+  'История, рассказанная через объектив',
+];
+
+const generateCommentId = createIdGenerator();
+
+const createMessage = () => Array.from({length: getRandomInteger(1, 2)}, () => getRandomArrayElement(MESSAGES)).join(' ');
+
+const createComment = () => ({
+  id: generateCommentId(),
+  avatar: `img/avatar-${getRandomInteger(AVATAR_MIN, AVATAR_MAX)}.svg`,
+  message: createMessage(),
+  name: getRandomArrayElement(NAMES),
+});
+
+const createPhotoCard = (index) => ({
+  id: index,
+  url: `photos/${index}.jpg`,
+  description: getRandomArrayElement(DESCRIPTIONS),
+  likes: getRandomInteger(LIKES_MIN, LIKES_MAX),
+  comments: Array.from({length: getRandomInteger(COMMENTS_MIN, COMMENTS_MAX)}, createComment)
+});
+
+const getPhotoCards = () => Array.from(
+  {length: PICTURE_COUNT},
+  (_, index) => createPhotoCard(index + 1));
+
+export {getPhotoCards};
